requirements.txt
gdown
python-docx
langchain_google_genai
pdfplumber
dotenv
langgraph
google-api-python-client